Update adapters to use bottom cab
