+++ commons/src/main/kotlin/com/simplemobiletools/commons/adapters/FilepickerFavoritesAdapter.kt
@@ -8,6 +8,7 @@ import com.simplemobiletools.commons.R
 import com.simplemobiletools.commons.activities.BaseSimpleActivity
 import com.simplemobiletools.commons.extensions.getTextSize
 import com.simplemobiletools.commons.views.MyRecyclerView
+import com.simplemobiletools.commons.views.contextview.ContextView
 import kotlinx.android.synthetic.main.filepicker_favorite.view.*
 
 class FilepickerFavoritesAdapter(
@@ -35,7 +36,7 @@ class FilepickerFavoritesAdapter(
 
     override fun getItemCount() = paths.size
 
-    override fun prepareActionMode(menu: Menu) {}
+    override fun onContextViewCreated(view: ContextView) {}
 
     override fun actionItemPressed(id: Int) {}
 
@@ -46,8 +47,6 @@ class FilepickerFavoritesAdapter(
     override fun getItemKeyPosition(key: Int) = paths.indexOfFirst { it.hashCode() == key }
 
     override fun getItemSelectionKey(position: Int) = paths[position].hashCode()
-
-    override fun onActionModeCreated() {}
 
     override fun onActionModeDestroyed() {}
