+++ commons/src/main/kotlin/com/simplemobiletools/commons/adapters/MyRecyclerViewAdapter.kt
@@ -16,6 +16,10 @@ import com.simplemobiletools.commons.extensions.getAdjustedPrimaryColor
 import com.simplemobiletools.commons.extensions.getContrastColor
 import com.simplemobiletools.commons.interfaces.MyActionModeCallback
 import com.simplemobiletools.commons.views.MyRecyclerView
+import com.simplemobiletools.commons.views.contextview.ContextView
+import com.simplemobiletools.commons.views.contextview.ContextViewCallback
+import com.simplemobiletools.commons.views.contextview.ContextViewItem
+import com.simplemobiletools.commons.views.contextview.ContextViewPopup
 import java.util.*
 
 abstract class MyRecyclerViewAdapter(val activity: BaseSimpleActivity, val recyclerView: MyRecyclerView, val itemClick: (Any) -> Unit) :
@@ -32,13 +36,15 @@ abstract class MyRecyclerViewAdapter(val activity: BaseSimpleActivity, val recyc
     protected var selectedKeys = LinkedHashSet<Int>()
     protected var positionOffset = 0
     protected var actMode: ActionMode? = null
+    protected var contextCallback: ContextViewCallback? = null
+    protected var contextPopup: ContextViewPopup? = null
 
     private var actBarTextView: TextView? = null
     private var lastLongPressedItem = -1
 
     abstract fun getActionMenuId(): Int
 
-    abstract fun prepareActionMode(menu: Menu)
+    abstract fun onContextViewCreated(view: ContextView)
 
     abstract fun actionItemPressed(id: Int)
 
@@ -50,24 +56,27 @@ abstract class MyRecyclerViewAdapter(val activity: BaseSimpleActivity, val recyc
 
     abstract fun getItemKeyPosition(key: Int): Int
 
-    abstract fun onActionModeCreated()
-
     abstract fun onActionModeDestroyed()
 
     protected fun isOneItemSelected() = selectedKeys.size == 1
 
     init {
+        contextCallback = object : ContextViewCallback {
+            override fun onCreateContextView(view: ContextView) {
+                onContextViewCreated(view)
+            }
+
+            override fun onItemClicked(item: ContextViewItem) {
+                actionItemPressed(item.id)
+            }
+        }
+
         actModeCallback = object : MyActionModeCallback() {
             override fun onActionItemClicked(mode: ActionMode, item: MenuItem): Boolean {
-                actionItemPressed(item.itemId)
-                return true
+                return false
             }
 
             override fun onCreateActionMode(actionMode: ActionMode, menu: Menu?): Boolean {
-                if (getActionMenuId() == 0) {
-                    return true
-                }
-
                 isSelectable = true
                 actMode = actionMode
                 actBarTextView = layoutInflater.inflate(R.layout.actionbar_title, null) as TextView
@@ -80,18 +89,16 @@ abstract class MyRecyclerViewAdapter(val activity: BaseSimpleActivity, val recyc
                         selectAll()
                     }
                 }
-                activity.menuInflater.inflate(getActionMenuId(), menu)
-                onActionModeCreated()
                 return true
             }
 
             override fun onPrepareActionMode(actionMode: ActionMode, menu: Menu): Boolean {
-                prepareActionMode(menu)
-                return true
+                return false
             }
 
             override fun onDestroyActionMode(actionMode: ActionMode) {
                 isSelectable = false
+                contextPopup?.dismiss()
                 (selectedKeys.clone() as HashSet<Int>).forEach {
                     val position = getItemKeyPosition(it)
                     if (position != -1) {
@@ -143,6 +150,7 @@ abstract class MyRecyclerViewAdapter(val activity: BaseSimpleActivity, val recyc
         if (oldTitle != newTitle) {
             actBarTextView?.text = newTitle
             actMode?.invalidate()
+            contextPopup?.invalidate()
         }
     }
 
@@ -329,6 +337,7 @@ abstract class MyRecyclerViewAdapter(val activity: BaseSimpleActivity, val recyc
             val currentPosition = adapterPosition - positionOffset
             if (!actModeCallback.isSelectable) {
                 activity.startSupportActionMode(actModeCallback)
+                contextPopup = ContextViewPopup(activity, getActionMenuId()).also { it.show(contextCallback) }
             }
 
             toggleItemSelection(true, currentPosition, true)
