+++ samples/src/main/kotlin/com/simplemobiletools/commons/samples/activities/StringsAdapter.kt
@@ -1,6 +1,5 @@
 package com.simplemobiletools.commons.samples.activities
 
-import android.view.Menu
 import android.view.MotionEvent
 import android.view.View
 import android.view.ViewGroup
@@ -15,6 +14,7 @@ import com.simplemobiletools.commons.interfaces.ItemTouchHelperContract
 import com.simplemobiletools.commons.interfaces.StartReorderDragListener
 import com.simplemobiletools.commons.samples.R
 import com.simplemobiletools.commons.views.MyRecyclerView
+import com.simplemobiletools.commons.views.contextview.ContextView
 import kotlinx.android.synthetic.main.list_item.view.*
 import java.util.*
 
@@ -41,7 +41,7 @@ class StringsAdapter(
 
     override fun getActionMenuId() = R.menu.cab_delete_only
 
-    override fun prepareActionMode(menu: Menu) {}
+    override fun onContextViewCreated(view: ContextView) {}
 
     override fun actionItemPressed(id: Int) {
         if (selectedKeys.isEmpty()) {
@@ -62,8 +62,6 @@ class StringsAdapter(
     override fun getItemKeyPosition(key: Int) = strings.indexOfFirst { it.hashCode() == key }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int) = createViewHolder(R.layout.list_item, parent)
-
-    override fun onActionModeCreated() {}
 
     override fun onActionModeDestroyed() {
         if (isChangingOrder) {
